Moves apply to backlog option
